Add retrieving storage size to payload storage trait (#5523)

+++ lib/segment/src/common/rocksdb_wrapper.rs
@@ -276,11 +276,11 @@ impl DatabaseColumnWrapper {
     /// Get the size of the storage in bytes
     ///
     /// The size of this column family in bytes, which is equal to the sum of the file size of its "levels"
-    pub fn get_storage_size_bytes(&self) -> OperationResult<u64> {
+    pub fn get_storage_size_bytes(&self) -> OperationResult<usize> {
         let db = self.database.read();
         let cf_handle = self.get_column_family(&db)?;
         let size = db.get_column_family_metadata_cf(cf_handle).size;
-        Ok(size)
+        Ok(size as usize)
     }
 }
